quiz-app/requirements.txt
flask==3.1.1
openai==2.32.0
apscheduler==3.11.0
python-dotenv==1.1.0
